Added the gscore function and hscore function as parameters, and changed the struct that astar takes in so it doesn't assume position.
I'm not yet sure if it's the best solution so if there's anybody around who can think of a better one that'd be nice.
Astar takes in data. Prior to this commit, it looked like this:

{
	pos pos;
	char blocked;
	struct node * breadcrumb;
	long hscore;
	long gscore;
	long fscore;
	struct node **connected_nodes;
	unsigned num_nodes;
	unsigned qid;
}
But it doesn't actually do anything with "pos", that's just so that after its' done it's job, following the breadcrumbs is actually meaningful.

Although it works, I'd like to use astar for things without a 2d position.
My implementation of it isn't particularly good for a 2d grid(If a 2d grid were assumed, for instance, no node would have to point to any other because you can discover them by adjacent position)

For this reason, I've tried to move away from that. Now, I have a void* for any userdata(In this case, position), and astar_data full of everything else:
typedef struct node {
	void* data;
	astar_data a_data;
} node;

All code has been changed to accomodate this.